refactor(lowering): extract ld reg-mem helpers
